backend/requirements.txt
# FastAPI Framework
fastapi
uvicorn[standard]

# Database
supabase>=2.4.0

# AI / LLM
openai
google-generativeai
tiktoken

# HTTP & Async
httpx
python-multipart

# Document Processing (RAG)
python-docx
PyPDF2

# Configuration
python-dotenv
pydantic
pydantic-settings
